Take into account the user's creation date
If the user was created sufficiently recently but has not yet
authenticated then we should not disable the user.

Note that we also determine the Kerberos domain in order to pull the
creation and last authentication timestamps directly from LDAP.  This
is because:
- The user creation timestamp is _only_ available from LDAP.
- The last authentication timestamp from LDAP is the most recent
across all IPA servers.  When we previously obtained the last
authentication timestamp via ipa user-status we got the most recent
timestamp from each server.